+++ visionforge-solid/src/commonMain/kotlin/space/kscience/visionforge/solid/specifications/AxesScheme.kt
@@ -7,24 +7,24 @@ import space.kscience.dataforge.meta.descriptors.MetaDescriptor
 import space.kscience.dataforge.meta.descriptors.value
 import space.kscience.dataforge.meta.double
 
-public class Axes : Scheme() {
+public class AxesScheme : Scheme() {
     public var visible: Boolean by boolean(false)
     public var size: Double by double(AXIS_SIZE)
     public var width: Double by double(AXIS_WIDTH)
 
-    public companion object : SchemeSpec<Axes>(::Axes) {
+    public companion object : SchemeSpec<AxesScheme>(::AxesScheme) {
         public const val AXIS_SIZE: Double = 1000.0
         public const val AXIS_WIDTH: Double = 3.0
 
         override val descriptor: MetaDescriptor by lazy {
             MetaDescriptor {
-                value(Axes::visible){
+                value(AxesScheme::visible){
                     default(false)
                 }
-                value(Axes::size){
+                value(AxesScheme::size){
                     default(AXIS_SIZE)
                 }
-                value(Axes::width){
+                value(AxesScheme::width){
                     default(AXIS_WIDTH)
                 }
             }