+++ visionforge-solid/src/commonMain/kotlin/space/kscience/visionforge/solid/specifications/ControlsScheme.kt
@@ -3,6 +3,7 @@ package space.kscience.visionforge.solid.specifications
 import space.kscience.dataforge.meta.Scheme
 import space.kscience.dataforge.meta.SchemeSpec
 
-public class Light : Scheme() {
-    public companion object : SchemeSpec<Light>(::Light)
+
+public class ControlsScheme : Scheme() {
+    public companion object : SchemeSpec<ControlsScheme>(::ControlsScheme)
 }