+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 
 ### Changed
+- Naming of Canvas3D options
+- Lights are added to the scene instead of 3D options 
 
 ### Deprecated